+++ test/unit/ModuleInterfaceTest.php
@@ -3,7 +3,8 @@
 namespace Dhii\Modular\UnitTest\Module;
 
 use Dhii\Modular\Module\ModuleInterface as TestSubject;
-use Xpmock\TestCase;
+use PHPUnit\Framework\MockObject\MockObject as MockObject;
+use PHPUnit\Framework\TestCase;
 
 /**
  * Tests {@see TestSubject}.
@@ -13,26 +14,16 @@ use Xpmock\TestCase;
 class ModuleInterfaceTest extends TestCase
 {
     /**
-     * The name of the test subject.
-     *
-     * @since [*next-version*]
-     */
-    const TEST_SUBJECT_CLASSNAME = 'Dhii\Modular\Module\ModuleInterface';
-
-    /**
      * Creates a new instance of the test subject.
      *
      * @since [*next-version*]
      *
-     * @return TestSubject
+     * @return TestSubject&MockObject
      */
     public function createInstance()
     {
-        $mock = $this->mock(static::TEST_SUBJECT_CLASSNAME)
-            ->getKey()
-            ->setup()
-            ->run()
-            ->new();
+        $mock = $this->getMockBuilder(TestSubject::class)
+            ->getMock();
 
         return $mock;
     }
@@ -47,15 +38,9 @@ class ModuleInterfaceTest extends TestCase
         $subject = $this->createInstance();
 
         $this->assertInstanceOf(
-            static::TEST_SUBJECT_CLASSNAME,
+            TestSubject::class,
             $subject,
             'A valid instance of the test subject could not be created'
-        );
-
-        $this->assertInstanceOf(
-            'Dhii\Data\KeyAwareInterface',
-            $subject,
-            'Subject does not implement a required interface'
         );
     }
 }