+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [[*next-version*]] - YYYY-MM-DD
+### Changed
+- Module `setup()` now returns a `ServiceProviderInterface` instance.
+- Module `run()` now requires the `ContainerInterface` argument.
+- Modules are no longer key-aware.
+
+### Removed
+- `DependenciesAwareInterface` has been removed.
+- `ModuleFactoryInterface` has been removed.
+- `ModuleKeyAwareInterface` has been removed.
 
 ## [0.1] - 2019-11-05
 Stable release